+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + Add Instant to TimeSource
 + Add Refresh Token Service
 + Rename repository test package
++ Add Authentication util class
 
 ## 0.11.0
 + Remove unused imports